remodeling the code a bit to implement react-router-dom
