+++ src/index.jsx
@@ -2,14 +2,12 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { HelmetProvider } from 'react-helmet-async';
 import App from "./app.jsx";
-import AltTextOrgClient from "./alt-text-org";
 
-const altTextOrgClient = new AltTextOrgClient()
 
 ReactDOM.render(
     <React.StrictMode>
         <HelmetProvider>
-            <App altTextOrgClient={altTextOrgClient} />
+            <App/>
         </HelmetProvider>
     </React.StrictMode>,
     document.getElementById("root")
